requirements.txt
amqp==5.1.1
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
asyncio==3.4.3
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
exceptiongroup==1.1.1
Flask==2.2.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
outcome==1.2.0
pocketbase==0.8.0
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pyee==8.2.2
pyppeteer==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.2
pytz==2023.3
rfc3986==1.5.0
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.15.0
